+++ temp_env.sh
@@ -1,1 +1,0 @@
-export SUPABASE_DB_URL="postgres://postgres:!77Kukusha77!@db.jxyavtdmcloxnhuavokc.supabase.co:5432/postgres"